requirements.txt
requests>=2.28.0
fastapi>=0.78.0
uvicorn>=0.17.6
pycryptodome>=3.14.1
